LPD-85688 Moves classes to osb-faro-provisioning-client, so we can reuse the logic
