+++ modules/dxp/apps/osb/osb-faro/osb-faro-provisioning-client/src/main/java/com/liferay/osb/faro/provisioning/client/constants/FaroSubscriptionConstants.java
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.osb.faro.web.internal.constants;
+package com.liferay.osb.faro.provisioning.client.constants;
 
-import com.liferay.osb.faro.provisioning.client.constants.ProductConstants;
-import com.liferay.osb.faro.web.internal.subscription.FaroSubscriptionPlan;
+import com.liferay.osb.faro.provisioning.client.subscription.FaroSubscriptionPlan;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 
 import java.util.Map;
